backend/requirements.txt
# 项目框架，因为其方便构建API
fastapi  
uvicorn[standard]   

# 设置等需求
pydantic
pydantic-settings

# 网络交互
requests

# 文件处理
python-docx  # 用于读取 .docx 文件
PyPDF2       # 用于读取 .pdf 文件
python-multipart # FastAPI需要此库来处理文件上传
rarfile # 对rar文件的处理

# 数据库
sqlalchemy
aiosqlite # 对sqlite的异步处理

# 查重检测
scikit-learn